feat: 添加 GTest 可选支持 + 跨平台兼容性改进
- 顶层 CMakeLists.txt 添加 ENABLE_TESTS 选项
- GTest 支持：优先使用系统安装，失败时 FetchContent
- Project1/Project2 测试子目录改为条件启用
- 修复 CTest 启用时 'test' 目标名冲突
- 修复 LootSpawnSystem.h 中 RAND_MAX 隐式转换警告 (MSVC 兼容)
- 添加 COMPATIBILITY_REPORT.md 跨平台兼容性报告
- 添加 build_test.sh 自动化测试脚本

跨平台支持:
✅ macOS (AppleClang) - 已验证
⚠️ Windows MSVC - 配置就绪待验证
⚠️ Linux - 配置就绪待验证

测试:
- macOS 构建测试通过
- loot_pipeline_test 单元测试通过 (1/1)

+++ Project2/include/systems/LootSpawnSystem.h
@@ -58,8 +58,8 @@ public:
                     Entity lootEntity = ecs.create();
                     
                     // 【修复】添加随机抖动 (±20 像素)
-                    float jitterX = (static_cast<float>(std::rand()) / RAND_MAX) * 40.0f - 20.0f;
-                    float jitterY = (static_cast<float>(std::rand()) / RAND_MAX) * 40.0f - 20.0f;
+                    float jitterX = (static_cast<float>(std::rand()) / static_cast<float>(RAND_MAX)) * 40.0f - 20.0f;
+                    float jitterY = (static_cast<float>(std::rand()) / static_cast<float>(RAND_MAX)) * 40.0f - 20.0f;
                     
                     // 【修复】Transform：必须使用死者的 position + 抖动！
                     transforms.add(lootEntity, {
@@ -68,8 +68,8 @@ public:
                         .rotation = 0.0f,
                         // ← 新增：爆米花抛射效果（更快的初速度）
                         .velocity = {
-                            (static_cast<float>(std::rand()) / RAND_MAX) * 120.0f - 60.0f,  // x: -60 ~ 60
-                            (static_cast<float>(std::rand()) / RAND_MAX) * -120.0f - 80.0f  // y: -200 ~ -80 (向上更快)
+                            (static_cast<float>(std::rand()) / static_cast<float>(RAND_MAX)) * 120.0f - 60.0f,  // x: -60 ~ 60
+                            (static_cast<float>(std::rand()) / static_cast<float>(RAND_MAX)) * -120.0f - 80.0f  // y: -200 ~ -80 (向上更快)
                         }
                     });
